Fixed reset of repeated mentions counter
These changes were prompted by a failing test (A9). Test A9 was failing
with the "Found too many repeated mentions (> 10) in the response, so
refusing to score. Please fix the output." error. There are not > 10
repeated mentions in the response for Test A9, however, the use of
$main::repeated_mentions in the code resulted in the repeated mentions
being counted once for every scoring metric used. Since there are 5
scoring metrics used in Test A9, 5 times the number of repeated
mentions were recorded which easily surpasses the threshold of 10. The
likely commit that introduced this issue is:
1d87110f168774cd38792c65c0d05ef8e42c9589. The fix proposed here does
not use $main:: and instead declares the $repeated_mentions counter
outside of the Score() subroutine.